+++ src/renderer/src/components/settings/accounts-search.ts
@@ -193,16 +193,16 @@ export const getAccountsMiniMaxSearchEntries = createLocalizedCatalog(() => [
 
 export const getAccountsGrokSearchEntries = createLocalizedCatalog(() => [
   {
-    title: translate('auto.components.settings.accounts.search.grokTitle', 'Grok (xAI) Usage'),
+    title: translate('auto.components.settings.accounts.search.f4a8c2e1b7', 'Grok (xAI) Usage'),
     description: translate(
-      'auto.components.settings.accounts.search.grokDesc',
+      'auto.components.settings.accounts.search.e3b7d1f9a2',
       'OAuth sign-in via Grok CLI (grok login) for weekly credit usage.'
     ),
     keywords: [
-      ...translateSearchKeyword('auto.components.settings.accounts.search.grokKw', 'grok'),
-      ...translateSearchKeyword('auto.components.settings.accounts.search.xaiKw', 'xai'),
-      ...translateSearchKeyword('auto.components.settings.accounts.search.oauthKw', 'oauth'),
-      ...translateSearchKeyword('auto.components.settings.accounts.search.loginKw', 'login'),
+      ...translateSearchKeyword('auto.components.settings.accounts.search.d2c6a0e8f1', 'grok'),
+      ...translateSearchKeyword('auto.components.settings.accounts.search.c1b5f9d7e0', 'xai'),
+      ...translateSearchKeyword('auto.components.settings.accounts.search.b0a4e8c6d9', 'oauth'),
+      ...translateSearchKeyword('auto.components.settings.accounts.search.a9f3d7b5c8', 'login'),
       ...translateSearchKeyword(
         'auto.components.settings.accounts.search.e949b08ffb',
         'rate limit'
